+++ src/main/java/joshie/harvest/animals/block/BlockTray.java
@@ -6,7 +6,6 @@ import joshie.harvest.animals.tile.TileNest;
 import joshie.harvest.api.HFApi;
 import joshie.harvest.api.animals.*;
 import joshie.harvest.api.core.Size;
-import joshie.harvest.core.HFCore;
 import joshie.harvest.core.achievements.HFAchievements;
 import joshie.harvest.core.base.block.BlockHFEnum;
 import joshie.harvest.core.base.tile.TileFillable;
@@ -16,8 +15,6 @@ import joshie.harvest.core.lib.CreativeSort;
 import net.minecraft.block.SoundType;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.item.EntityItem;
 import net.minecraft.entity.passive.EntityAnimal;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.init.SoundEvents;
@@ -33,14 +30,13 @@ import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 
 import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
-import java.util.List;
 import java.util.Locale;
 
 import static joshie.harvest.animals.block.BlockTray.Tray.FEEDER_EMPTY;
 import static joshie.harvest.animals.block.BlockTray.Tray.NEST_EMPTY;
 
 public class BlockTray extends BlockHFEnum<BlockTray, Tray> implements IAnimalFeeder, INest {
+    private static final AxisAlignedBB NEST_COLLISION = new AxisAlignedBB(0.15D, 0D, 0.15D, 0.85D, 0.2D, 0.85D);
     private static final AxisAlignedBB NEST_AABB = new AxisAlignedBB(0.15D, 0D, 0.15D, 0.85D, 0.35D, 0.85D);
     private static final AxisAlignedBB FEEDER_AABB = new AxisAlignedBB(0.0D, 0D, 0.0D, 1.0D, 0.075D, 1.0D);
 
@@ -68,15 +64,15 @@ public class BlockTray extends BlockHFEnum<BlockTray, Tray> implements IAnimalFe
         return "axe";
     }
 
-    @SuppressWarnings("deprecation")
     @Override
-    public void addCollisionBoxToList(IBlockState state, @Nonnull World worldIn, @Nonnull BlockPos pos, @Nonnull AxisAlignedBB entityBox, @Nonnull List<AxisAlignedBB> collidingBoxes, @Nullable Entity entityIn) {
-        if (entityIn instanceof EntityPlayer || entityIn instanceof EntityItem) super.addCollisionBoxToList(state, worldIn, pos, entityBox, collidingBoxes, entityIn);
-        else addCollisionBoxToList(pos, entityBox, collidingBoxes, HFCore.FENCE_COLLISION);
+    @SuppressWarnings("deprecation, unchecked")
+    public AxisAlignedBB getCollisionBoundingBox(IBlockState state, @Nonnull World world, @Nonnull BlockPos pos) {
+        return getEnumFromState(state).isFeeder() ? super.getCollisionBoundingBox(state, world, pos) : NEST_COLLISION;
     }
 
     @SuppressWarnings("deprecation")
     @Override
+    @Nonnull
     public AxisAlignedBB getBoundingBox(IBlockState state, IBlockAccess world, BlockPos pos) {
         return getEnumFromState(state).isFeeder() ? FEEDER_AABB : NEST_AABB;
     }
@@ -117,15 +113,15 @@ public class BlockTray extends BlockHFEnum<BlockTray, Tray> implements IAnimalFe
 
     @Override
     public boolean feedAnimal(AnimalStats stats, World world, BlockPos pos, IBlockState state, boolean simulate) {
-        if (getEnumFromState(state).isFeeder()) {
-            if (HFApi.animals.canAnimalEatFoodType(stats, AnimalFoodType.SEED)) {
-                TileFeeder feeder = ((TileFeeder) world.getTileEntity(pos));
-                if (feeder != null && feeder.getFillAmount() > 0) {
-                    if (simulate) return true;
+        if (getEnumFromState(state).isFeeder() && HFApi.animals.canAnimalEatFoodType(stats, AnimalFoodType.SEED)) {
+            TileFeeder feeder = ((TileFeeder) world.getTileEntity(pos));
+            if (feeder != null && feeder.getFillAmount() > 0) {
+                if (!simulate) {
                     feeder.adjustFill(-1);
                     stats.performAction(world, null, null, AnimalAction.FEED);
-                    return true;
                 }
+
+                return true;
             }
         }
 
@@ -133,10 +129,15 @@ public class BlockTray extends BlockHFEnum<BlockTray, Tray> implements IAnimalFe
     }
 
     @Override
-    public boolean layEgg(AnimalStats stats, World world, BlockPos pos, IBlockState state) {
+    public boolean isNest(AnimalStats stats, World world, BlockPos pos, IBlockState state) {
+        return getEnumFromState(state.getActualState(world, pos)) == NEST_EMPTY;
+    }
+
+    @Override
+    public void layEgg(AnimalStats stats, World world, BlockPos pos, IBlockState state) {
         EntityPlayer player = stats.getOwner();
         TileEntity tile = world.getTileEntity(pos);
-        if (player != null && tile instanceof TileNest) {
+        if (player != null && stats.getAnimal() != null && tile instanceof TileNest) {
             if (!world.isRemote) {
                 ((TileNest) tile).setDrop(EntityHelper.getEntityUUID(stats.getAnimal()), stats.getType().getProduct(player, stats));
                 stats.setProduced(1); //Product one egg
@@ -144,10 +145,7 @@ public class BlockTray extends BlockHFEnum<BlockTray, Tray> implements IAnimalFe
 
             EntityAnimal animal = stats.getAnimal();
             animal.playSound(SoundEvents.ENTITY_CHICKEN_EGG, 1.0F, (animal.worldObj.rand.nextFloat() - animal.worldObj.rand.nextFloat()) * 0.2F + 1.0F);
-            return true;
         }
-
-        return false;
     }
 
     @Override
@@ -156,13 +154,15 @@ public class BlockTray extends BlockHFEnum<BlockTray, Tray> implements IAnimalFe
     }
 
     @Override
-    public TileEntity createTileEntity(World world, IBlockState state) {
+    @Nonnull
+    public TileEntity createTileEntity(@Nonnull World world, @Nonnull IBlockState state) {
         return getEnumFromState(state).isFeeder() ? new TileFeeder() : new TileNest();
     }
 
     @SuppressWarnings("deprecation")
     @Override
-    public IBlockState getActualState(IBlockState state, IBlockAccess world, BlockPos pos) {
+    @Nonnull
+    public IBlockState getActualState(@Nonnull IBlockState state, IBlockAccess world, BlockPos pos) {
         TileEntity tile = world.getTileEntity(pos);
         if (tile instanceof TileFeeder) {
             boolean isFilled = ((TileFeeder)tile).getFillAmount() > 0;