+++ src/main/java/joshie/harvest/animals/entity/ai/EntityAIEat.java
@@ -12,6 +12,7 @@ import javax.annotation.Nonnull;
 public class EntityAIEat extends EntityAIAnimal {
     private boolean isReadyToEat;
     private int currentTask;
+    private int eatTimer;
 
     public EntityAIEat(EntityAnimal animal) {
         super(animal);
@@ -36,24 +37,28 @@ public class EntityAIEat extends EntityAIAnimal {
     @Override
     public void updateTask() {
         super.updateTask();
-        animal.getLookHelper().setLookPosition((double)destinationBlock.getX() + 0.5D, (double)(destinationBlock.getY() + 1), (double)destinationBlock.getZ() + 0.5D, 10.0F, (float)animal.getVerticalFaceSpeed());
         World world = animal.worldObj;
-        BlockPos blockpos = destinationBlock.up();
-        if (animal.getDistanceSq(blockpos) <= 3D) {
-            IBlockState iblockstate = world.getBlockState(blockpos);
-            if (currentTask == 1 && isEdible(iblockstate)) {
-                eat(blockpos, iblockstate);
-            }
+        if (animal.getDistance(destinationBlock.getX(), destinationBlock.getY(), destinationBlock.getZ()) <= 1D) {
+            animal.getLookHelper().setLookPosition((double)destinationBlock.getX() + 0.5D, (double)(destinationBlock.getY()), (double)destinationBlock.getZ() + 0.5D, 10.0F, (float)animal.getVerticalFaceSpeed());
+            if (eatTimer == 0) eatTimer = 20;
+            else eatTimer--;
 
-            currentTask = -1;
-            runDelay = 0;
+            if (eatTimer <= 0) {
+                IBlockState iblockstate = world.getBlockState(destinationBlock);
+                if (currentTask == 1 && isEdible(destinationBlock, iblockstate)) {
+                    eat(destinationBlock, iblockstate);
+                }
+
+                currentTask = -1;
+                runDelay = 10;
+            }
         }
     }
 
     @Override
     protected boolean shouldMoveTo(@Nonnull World worldIn, @Nonnull BlockPos pos) {
-        IBlockState iblockstate = worldIn.getBlockState(pos.up());
-        if (animal.getDistanceSq(pos.up()) <= 32D && isEdible(iblockstate)) {
+        IBlockState iblockstate = worldIn.getBlockState(pos);
+        if (animal.getDistance(pos.getX(), pos.getY(), pos.getZ()) <= 64D && isEdible(pos, iblockstate)) {
             if (isReadyToEat && (currentTask == 1 || currentTask < 0)) {
                 currentTask = 1;
                 return true;
@@ -63,8 +68,8 @@ public class EntityAIEat extends EntityAIAnimal {
         return false;
     }
 
-    protected boolean isEdible(IBlockState iblockstate) {
-        return iblockstate.getBlock() instanceof IAnimalFeeder;
+    boolean isEdible(BlockPos pos, IBlockState state) {
+        return state.getBlock() instanceof IAnimalFeeder && ((IAnimalFeeder) state.getBlock()).feedAnimal(getStats(), animal.worldObj, pos, state, true);
     }
 
     protected void eat(BlockPos pos, IBlockState state) {